#  Use the same code-flow regardless of metrics being enabled