--- a/internal/storage/bucket_handle.go
+++ b/internal/storage/bucket_handle.go
@@ -95,8 +95,7 @@ func (bh *bucketHandle) NewReaderWithReadHandle(
 	}
 
 	// NewRangeReader creates a "storage.Reader" object which is also io.ReadCloser since it contains both Read() and Close() methods present in io.ReadCloser interface.
-	var storageReader *storage.Reader
-	storageReader, err = obj.NewRangeReader(ctx, start, length)
+	storageReader, err := obj.NewRangeReader(ctx, start, length)
 	if err == nil {
 		reader = newGCSFullReadCloser(storageReader)
 	}